feat: implement project generation command with framework validation
