+++ pyproject.toml
@@ -58,7 +58,6 @@ build-backend = "uv_build"
 mypy_path = ["src"]
 exclude = [
 	"src/scaffoldr/templates/.*",
-	"src/scaffoldr/templates",
 ]
 namespace_packages = true
 explicit_package_bases = true
