+++ src/scaffoldr/cli/__init__.py
@@ -1,0 +1,3 @@
+from .generate import gen
+
+__all__ = ["gen"]
