+++ src/scaffoldr/core/constants/__init__.py
@@ -1,4 +1,8 @@
 from .art import ascii_art
-from .const import console
+from .const import Frameworks, console
 
-__all__ = ["ascii_art", "console"]
+__all__ = [
+	"Frameworks",
+	"ascii_art",
+	"console",
+]
