+++ src/scaffoldr/core/constants/art.py
@@ -1,8 +1,8 @@
 ascii_art = [
-    list("███████╗ ██████╗ █████╗ ███████╗███████╗ ██████╗ ██╗     ██████╗ ██████╗ "),
-    list("██╔════╝██╔════╝██╔══██╗██╔════╝██╔════╝██╔═══██╗██║     ██╔══██╗██╔══██╗"),
-    list("███████╗██║     ███████║█████╗  █████╗  ██║   ██║██║     ██║  ██║██████╔╝"),
-    list("╚════██║██║     ██╔══██║██╔══╝  ██╔══╝  ██║   ██║██║     ██║  ██║██╔══██╗"),
-    list("███████║╚██████╗██║  ██║██║     ██║     ╚██████╔╝███████╗██████╔╝██║  ██║"),
-    list("╚══════╝ ╚═════╝╚═╝  ╚═╝╚═╝     ╚═╝      ╚═════╝ ╚══════╝╚═════╝ ╚═╝  ╚═╝"),
+	list("███████╗ ██████╗ █████╗ ███████╗███████╗ ██████╗ ██╗     ██████╗ ██████╗ "),
+	list("██╔════╝██╔════╝██╔══██╗██╔════╝██╔════╝██╔═══██╗██║     ██╔══██╗██╔══██╗"),
+	list("███████╗██║     ███████║█████╗  █████╗  ██║   ██║██║     ██║  ██║██████╔╝"),
+	list("╚════██║██║     ██╔══██║██╔══╝  ██╔══╝  ██║   ██║██║     ██║  ██║██╔══██╗"),
+	list("███████║╚██████╗██║  ██║██║     ██║     ╚██████╔╝███████╗██████╔╝██║  ██║"),
+	list("╚══════╝ ╚═════╝╚═╝  ╚═╝╚═╝     ╚═╝      ╚═════╝ ╚══════╝╚═════╝ ╚═╝  ╚═╝"),
 ]
